Finish base structure for second question
Created the basic interaction for the second question.  Added the
new intent to the interaction model along with sample utterances.
Took the opportunity to add more sample utterances to the first
question's intent to increase usability.  Note, this still does
not prepare the user's response for later processing.  This is
strictly the interaction without long term storage of responses.