Set up auto incrementing Plan of Action id numbers
The initial plan to connect to S3 to store and increment id numbers
for each plan of action didn't quite pan out.  Instead, persistence
was achieved using the DynamoDB persistence adapter.

Current state is a simple system for getting the next id number from
a DynamoDBtable, assigning it to the plan of action, incrementing the
number, and saving it back to the DynamoDB table.  This update serves
as a proof of concept that will likely be expeanded upon later.

The auto-generated id number and all user inputs are successfully saved
into the correct DynamoDB table.