Final confirmation: Yes option
Adjusted the code to read back the completed plan of action at the end
of the process for one final confirmation of user entry.  This code
implements the successful use case of the user saying yes, which will
trigger the skill to save the data to DynamoDB.  A little off topic,
but some slight changes were made to the interaction model to try
to eliminate some awkward responses from Alexa. (Things like her saying
'was was' in a response)  Further debugging and developing of the interaction
model will be required.