Final confirmation: no response
This commit sets up the code to accept a no response from the user
if they answer no to the final confirmation of their data entry. The
code is in place to allow the user to start the process from the
beginning of the process.

It's a little unstable for some reason and works intermittently for
no apparent reason.  Each deployment seems to either make it work or
break it without changing anything in the code.  Will debug further
by next commit.