Started finalizing fallback intent
Up to this point, the fallback intent was only implemented for the
key user entry points of the skill.  This update implements the fallback
intent for the first few yes/no confirmations in the skill.  Also discovered
an instance where the user could answer no to a confirmation and there was
an error response.  Fixed this instance.