Retrieve account status
To simulate getting the status of an account to see what is needed to reinstate
it, I set up a separate DynamoDB table to hold sample account data in the form
of a simple status code.

0 = Good standing
1 = POA required
2 = Self-reinstatement eligible

This update implements the helper method to retrieve the account data from DynamoDB.

Moving forward, if we implement the case of multiple infractions we can just add more
items to the data base and add attributes for more specific infraction cases.