Refactor skill invocation
This update refactors the beginning of the skill to use the retrieved
account status code to direct the user to the appropriate workflow.

TODO: Change the invocation of the skill.  'Make a plan of action' isn't
accurate anymore.  Something along the lines of 'Fix my account' or
'Check my account status' would be more appropriate.