Completed the self-reinstatement sequence
The skill can now guide the user through the complete self-reinstatment
process and will update the account status to statusCode 0 (all clear)
when the process is complete.  If the user says no at any point in the
process (verbal equivalent of not clicking the necessary check box) the
skill will explain what is necessary for self-reinstatement and end the
skill.

Further developement of the Help intent and Fallback intent will be needed
to better support the expanding functionality of the skill.