Completed task 99
This commit completes the core functionality of the reply feature
which allows the user to add additional information to an already submitted
plan of action.  The skill successfully retrieves the submitted plan of
action id, prompts the user for the additional data, and appends the
data to the dynamodb entry for that plan of action.