Completed task 100
Updated the help and fallback intents to include prompts for the
reply functionality.  Tested on Alexa device with expected results.
Text input on the Alexa simulator in the dev console did not always
trigger the fallback as intended, disregarding this behavior for now
as voice input on both the simulator and an Alexa device worked
as expected.