Task 70_71-DialogModel
Set up the dialog model in an attempt to simplify the structure of
the code.  This should allow for much more readable logic code, with
the conversation aspect taken care of in the interaction model rather
than in the business logic.

Will continue to fine tune the model while refactoring the logic in
task 70_72.