70_112 Plan of Action Dialog Model
This updates refactors the code to rely on the dialog model for a
more natural user experience.  Using the dialog model removes the need
for 'carrier phrases' to trigger specific intents to capture information.
Rather, through slot elicitation the user can say whatever they want
and the skill will capture it.  Confirmation of data entry is also handled
by the dialog model.  The Yes and No intents are only used now to confirm
the final plan of action before submission.  The help and fall back intents
are simplified because there are only a few interaction points to consider,
vs the pervious state of needing to account for every possible Yes/No confirmation
and intent trigger.