Task 125 - Reminders
Began work on implementing the reminders API to show proof of concept to
have Alexa automatically notify the user when there is an issue with their
account.  This update contains the initial, untested utility functions for
creating a reminder.  Next update will integrate these functions with the
core skill code found in index.js.