Task 125 Complete
This update completes task 125.  The reminders API has been successfully
implemented, but it is not being utilized in any practical way. That is
to say, it successfully creates a reminder as expected, but the logic
to actually use this functionality for anything meaningful has not been
implemented yet.  The logic to itegrate the reminders API into the flow of
the skill will be covered in the next task (126).