Task 126
Started to implement a test scenario for the Reminders API to demonstrate
a proof of concept.  If the account status is all clear, the skill will
create a reminder for the user and set the account status to some non-all clear
state.  This update sets the reminder.  Next commit will include the status
change.