Task 126 Complete
This update completes task 126.  The reminders API will set a daily reminder
at 12PM for the user that informs them that their account needs
attention.  This functionality is triggered when there is nothing wrong with
the user's account, and will reset the status of the user's account to
needing a POA to fix.  This simulates an account going from being in good
standing, to violating a policy, to not being in good standing.