Minor Debug
Debugged the following:

Skill would not exit at the end of certain sequences as expected and would prompt the user
after the task was complete.

Retrieval of infraction information would sometimes result in 'undefined' due to asynchronous
behavior.

Setup proper response if the user responds with 'no' to the first question in the self-
reinstatement process.