Task 139
This updates adds a user prompt to be notified of any issues with the
account.  This functionality is more of a simulation of an idea rather
than actual functionality.  It uses the reminders api to simulate how
a future system could be implemented to allow Amazon to send audio notifications
to users via an Echo device to notify them of errors with their account.