Task 142
This update enables the skill to automatically send an email with
a confirmation message that is contextually accurate for the process
that has been completed.  That is to say, the POA, SR, and Reply processes
all generate different messages to be emailed.  Next update to finish this
task will be code cleanup.