Task 142 - Complete
Used this task as an opportunity to clean up the code.  Moved some
of the helper function from index.js to util.js, created a separate
file to hold the bulk of Alexa's responses as constants to make index.js
a little easier to navigate rather than having a bunch of strings
cluttering up the logic.