Task 122
This update includes screen support for the SR process and most
scenarios covered in the Yes intent, No intent, and cancel intent.
Note: Due to constraints in the Alexa platform, support for an
in-process Self-reinstatement is not possible at this time.  Screen
delegation and dialog delegation can not be combined as of this time.
To implement this, a major refactoring of the code would be necessary
and somewhat regressive to the work done so far.